[skip ci] Reving to 0.116.0 - HTTP PATCH and mergemap (#114)
* Adding PATCH support for updating Tenant Domain.
* Moved the PUT in '/v1/tenants/{tenantId}/domains' to '/v1/tenants/{tenantId}/domains/{domainId}' to better match convention.
* Added functions dbgDumpMapReflection(), dbgCompareMaps(),
mergeMaps(),mergeObjWithMap(). These functions are to be used for HTTP
PATCH with JSON API. The library we're using for JSONAPI does not have
support for partial update. "[Bulk creates and updates are not supported
yet.](https://godoc.org/github.com/google/jsonapi#UnmarshalManyPayload)

So the functions added will allow us to do partial updates without
writing a ton of 1-off code to support PATCH.
* Moved JSON API merge functions to utils.go. Added some PATCH support for
tenantUser, TenantIngestionProfile, TenantThresholdProfile.
* Adding datastore/test/utils_test.go to test GetDataIDFromFullID().
GetDataIDFromFullID() was updated with new logic to handle an edge case
with multiple underscores in the ID's string. If there were more than
the expected 2 underscores, it would truncate too much of the ID's
string.